pyproject.toml
[project]
name = "learn-ope"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.3.2",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.7.1",
]
